+++ app/src/main/java/com/example/tanify/data/response/artikel/ArtikelResponse.kt
@@ -1,4 +1,4 @@
-package com.example.tanify.data.response
+package com.example.tanify.data.response.artikel
 
 import com.google.gson.annotations.SerializedName
