+++ app/src/main/java/com/example/tanify/data/response/profile/UserProfilResponse.kt
@@ -1,0 +1,12 @@
+package com.example.tanify.data.response.profile
+
+import com.google.gson.annotations.SerializedName
+
+data class UserProfilResponse(
+    @SerializedName("nama")
+    val nama: String? = null,
+    @SerializedName("email")
+    val email: String? = null,
+    @SerializedName("photo")
+    val photo: String? = null,
+)
