+++ app/src/main/java/com/example/tanify/data/response/weather/CurrentWeatherResponse.kt
@@ -1,4 +1,4 @@
-package com.example.tanify.data.response
+package com.example.tanify.data.response.weather
 
 import com.google.gson.annotations.SerializedName
