+++ app/src/main/java/com/example/tanify/ui/bottomNav/forum/items/ItemFragmentForumAdapter.kt
@@ -1,0 +1,4 @@
+package com.example.tanify.ui.bottomNav.forum.items
+
+class ItemFragmentForumAdapter {
+}
